+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/) (or at lea
 
 ## [Unreleased]
 ### Added
-- Error Boundary
+- .htaccess now enables gzip compression
 - React Toastify to provide notification capabilities
 - Sentry for error tracking
 - Terms and conditions of use
@@ -15,6 +15,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/) (or at lea
 - CRUD components to container of Vanilla components, providing props about fields and paths
 ### Removed
 - Cards as messages (replace with toaster notification)
+- Errors introduced purposely to test Error Boundary in production
+
+## [0.6.0] - 2018-10-01
+### Added
+- Error Boundary
+- Errors to Register and Profile component to demonstrate Error Boundary in production
 
 ## [0.5.2] - 2018-10-01
 ### Fixed